Move redux container components
